chore: update taikungoclient and adapt to breaking API changes
Bump github.com/itera-io/taikungoclient to the latest nightly and adapt
to the Domain->Account rename and related signature changes:

- DomainId(...) -> AccountId(...) on list requests
- GetDomainId/GetDomainName -> GetAccountId/GetAccountName on RobotUsersListDto
- UserBriefDto -> UserListItemDto for offset-based user dropdown
- GetTotalCount() now returns int64 (projects pagination)
- Drop removed IsAdmin parameter from cloud credentials listing
- Update NewProjectsList call for its new signature

+++ projects.go
@@ -154,7 +154,7 @@ func listVirtualClusterProjects(ctx context.Context, client *taikungoclient.Clie
 			}
 		}
 
-		if int32(len(projectList.Data)) < pageSize || pageOffset+pageSize >= projectList.GetTotalCount() {
+		if int32(len(projectList.Data)) < pageSize || int64(pageOffset+pageSize) >= projectList.GetTotalCount() {
 			break
 		}
 	}
